debug

+++ serveur/main.js
@@ -88,16 +88,12 @@ app.get('/saves/get/:id', function (req, res) {
 
 app.post('/saves/post/:id', upload.any(), function (req, res) {
     console.log("REST upload file");
-    try
-    {
+    if (fs.existsSync(__dirname + '/tmp/' + req.files[0].filename)) {
         var id = req.params.id;
         var source = fs.createReadStream(__dirname + '/tmp/' + req.files[0].filename);
         var dest = fs.createWriteStream(__dirname + '/uploads/saves/' + id + '.npz');
         source.pipe(dest);
         fs.unlinkSync(__dirname + '/tmp/' + req.files[0].filename);
-    } catch (e)
-    {
-        console.log(e);
     }
     res.send(200);
 });
